Refactor playSong into multiple methods, utilizing youtube utilities
